% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion.R
\docType{data}
\name{deletion}
\alias{deletion}
\title{Drug prediction results on trimmed graphs after enrichment}
\format{
A list with 500 items, including 10\%, 20\%, 30\%, 40\%, 50\% failures,
and node and link failures.
}
\usage{
deletion
}
\description{
We we simulated 5 node failures and 5 link failures with 50
repeats in each condition to measure the robustness of the graph.
}
\examples{
library(DTSEAdata)
data("deletion")

# Get length
length(deletion)

}
\keyword{datasets}
