% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_indications.R
\docType{data}
\name{drug_indications}
\alias{drug_indications}
\title{The drug indications data used in the DTSEA paper}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32758 rows and 8 columns.
}
\usage{
drug_indications
}
\description{
We downloaded the drug indications data from ChEMBL, one of the
largest drug information databases in the world. In our provided data, we deleted the drugs without a DrugBank ID.
}
\details{
A data frame with 32758 rows and 8 columns:
\itemize{
\item \code{drugbank_id}: The DrugBank ID of the drugs.
\item \code{drug_name}: the drug name.
\item \code{max_phase}: refers to the phase a drug achieved in a clinical
trial. Phases are typically ordered from \strong{0} to \strong{4}:\cr
0. Preclinical research;
\enumerate{
\item Clinical phase 1;
\item Clinical phase 2;
\item Clinical phase 3;
\item Approved by FDA or EMCA.
}
\item \code{mesh_id}: the Mesh ID of the specific indications.
\item \code{mesh_heading}: the Mesh Heading of the specific indications.
\item \code{efo_id}: the EFO ID of the specific indications.
\item \code{efo_term}: the EFO Term of the specific indications.
\item \code{indication}: the text-based main indications of the drugs.
}
}
\examples{
library(DTSEAdata)
data("drug_indications")

# Get Aspirin
library(dplyr)
filter(drug_indications, drug_name == 'Acetylsalicylic acid') \%>\%
  select(-indication)

}
\references{
Gaulton A, Hersey A, Nowotka M, Bento AP, Chambers J, Mendez D, Mutowo P,
Atkinson F, Bellis LJ, Cibrián-Uhalte E, Davies M, Dedman N, Karlsson A,
Magariños MP, Overington JP, Papadatos G, Smit I, Leach AR. The ChEMBL
database in 2017. Nucleic Acids Res. 2017 Jan 4;45(D1):D945-D954. doi:
10.1093/nar/gkw1074. Epub 2016 Nov 28. PMID: 27899562; PMCID: PMC5210557.
}
\keyword{datasets}
