% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_predicted.R
\docType{data}
\name{drug_predicted}
\alias{drug_predicted}
\title{The predicted drug target data used in the DTSEA paper}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2401 rows and 10 columns.
}
\usage{
drug_predicted
}
\description{
We manually downloaded the predicted drug target data on the
\emph{Target Predictions} subsections of the Compound Report Card in each drug.
Also, you can run the prediction in your local machine using docker image
\code{chembl/mcp}.
}
\details{
A data frame with 2401 rows and 10 columns:
\itemize{
\item \code{drugbank_id}: the DrugBank ID
\item \code{drug_name}: the name of each drug
\item \code{gene_target}: the targets of drugs
\item \code{drugbank_id}: the DrugBank ID.
\item \code{drug_name}: the name of each drug.
\item \code{chembl_id}: the ChEMBL ID of each drug.
\item \code{uniport_id}: the UniPort ID of each protein.
\item \code{hgnc_symbol}: the HGNC symbol of each gene.
\item \code{target}: the ChEMBL ID of each target.
\item \code{seventy}: the 70\% confidence level of the prediction. The results are in
four levels: \strong{active}, \strong{inactive}, \strong{empty}, and \strong{both}.
The four levels represent the following:
\itemize{
\item \strong{Active}: the drug is predicted to interact with the target.
\item \strong{Inactive}: the drug is not predicted to interact with the target.
\item \strong{Empty}: the model can not predict the compound interacts with the target.
\item \strong{Both}: The model can not conclude the interaction.
}
\item \code{eighty}: the 80\% confidence level of the prediction. The results are in
four levels: \strong{active}, \strong{inactive}, \strong{empty}, and \strong{both}.
The four levels represent the following:
\itemize{
\item \strong{Active}: the drug is predicted to interact with the target.
\item \strong{Inactive}: the drug is not predicted to interact with the target.
\item \strong{Empty}: the model can not predict the compound interacts with the target.
\item \strong{Both}: The model can not conclude the interaction.
}
\item \code{ninty}: the 90\% confidence level of the prediction. The results are in
four levels: \strong{active}, \strong{inactive}, \strong{empty}, and \strong{both}.
The four levels represent the following:
\itemize{
\item \strong{Active}: the drug is predicted to interact with the target.
\item \strong{Inactive}: the drug is not predicted to interact with the target.
\item \strong{Empty}: the model can not predict the compound interacts with the target.
\item \strong{Both}: The model can not conclude the interaction.
}
\item \code{threshold}: activity threshold.
}
}
\examples{
library(DTSEAdata)
data("drug_predicted")

# Get the drug count
library(dplyr)
pull(drug_predicted, drug_name) \%>\%
  unique() \%>\%
  length()

}
\references{
Gaulton A, Hersey A, Nowotka M, Bento AP, Chambers J, Mendez D, Mutowo P,
Atkinson F, Bellis LJ, Cibrián-Uhalte E, Davies M, Dedman N, Karlsson A,
Magariños MP, Overington JP, Papadatos G, Smit I, Leach AR. The ChEMBL
database in 2017. Nucleic Acids Res. 2017 Jan 4;45(D1):D945-D954. doi:
10.1093/nar/gkw1074. Epub 2016 Nov 28. PMID: 27899562; PMCID: PMC5210557.

Bosc N, Atkinson F, Felix E, Gaulton A, Hersey A, Leach AR. Large scale
comparison of QSAR and conformal prediction methods and their applications in
drug discovery. J Cheminform. 2019 Jan 10;11(1):4. doi:
10.1186/s13321-018-0325-4. PMID: 30631996; PMCID: PMC6690068.
}
\keyword{datasets}
