% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTSEAdata-package.R
\name{DTSEAdata-package}
\alias{DTSEAdata-package}
\title{Data collections for Drug set enrichment analysis (DTSEA)}
\description{
This package provides supplementary data for \strong{DTSEA}. The data
package contains full disease-related gene list of GSE183071, full network
list, and other useful supplementary data for exploring DTSEA.

We also provide a list of all the supplementary data files.\tabular{ll}{
   Data name \tab Description \cr
   \code{deletion} \tab Drug prediction results on trimmed graphs after enrichment \cr
   \code{drug_indications} \tab The drug indications data from ChEMBL \cr
   \code{drug_predicted} \tab The drug target data predicted by ChEMBL \cr
   \code{drug_targets} \tab The combined drug target data \cr
   \code{expr_raw} \tab The raw expression profile of GSE164805 \cr
   \code{graph} \tab The human gene functional interaction network we used \cr
   \code{ncbi_list} \tab The COVID-19-related genes provided by NCBI \cr
}
}
\details{
DTSEAdata
}
