% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_raw.R
\docType{data}
\name{expr_raw}
\alias{expr_raw}
\title{The raw expression profile used in the DTSEA paper}
\format{
An object of class \code{data.frame} with 20180 rows and 15 columns.
}
\usage{
expr_raw
}
\description{
We provide the processed expression profile of GSE164805.
}
\details{
A data frame contains 20180 rows and 15 columns.
\itemize{
\item column: the patient label ID. IDs greater than \strong{0} are disease samples,
and less than \strong{0} are controls.
\item row: the gene symbol. The probes were converted into gene symbol and
aggregated the repeated ones by mean.
}
}
\examples{
library(DTSEAdata)
data("expr_raw")

# Get the data length
library(dplyr)
dim(expr_raw) \%>\%
  cat(sep = '*')

}
\references{
Zhang Q, Meng Y, Wang K, Zhang X, Chen W, Sheng J, Qiu Y, Diao H, Li L.
Inflammation and Antiviral Immune Response Associated With Severe Progression
of COVID-19. Front Immunol. 2021 Feb 18;12:631226. doi:
10.3389/fimmu.2021.631226. PMID: 33679778; PMCID: PMC7930228.
}
\keyword{datasets}
