% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{data}
\name{graph}
\alias{graph}
\title{The human gene functional interaction network object used in the DTSEA paper}
\format{
An \code{igraph} object with 221,353 edges and 12,836 nodes.
}
\usage{
graph
}
\description{
We extracted the gene functional interaction network from
multiple sources with experimental evidence and then integrated them.
}
\examples{
library(DTSEAdata)
data("graph")

# Get graph characteristics
library(igraph)
cat(length(V(graph)), 'nodes,', length(E(graph)), 'edges. \n')

}
\references{
Kanehisa M, Furumichi M, Sato Y, Ishiguro-Watanabe M, Tanabe M. KEGG:
integrating viruses and cellular organisms. Nucleic Acids Res. 2021 Jan 8;49
(D1):D545-D551. doi: 10.1093/nar/gkaa970. PMID: 33125081; PMCID: PMC7779016.
}
\keyword{datasets}
