% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_list.R
\docType{data}
\name{ncbi_list}
\alias{ncbi_list}
\title{The COVID-19-related genes provided by NCBI}
\format{
A \code{list} object with 190 gene symbols.
}
\usage{
ncbi_list
}
\description{
NCBI provided a COVID-19-related node list
(https://www.ncbi.nlm.nih.gov/gene/?term=coronavirus+related+\%5Bproperties\%5D).
The WayBackMachine did not snapshot the page on July 2022, so we manually
took a snapshot and obtained this gene list.
}
\examples{
library(DTSEAdata)
data("ncbi_list")

# Get vector size
length(ncbi_list)

}
\keyword{datasets}
